initial_setup/pyproject.toml
[tool.poetry]
name = "terraform-provider-creation"
version = "0.1.0"
description = "Scripts to set up the terraform private provider initially"
authors = ["Jack Mead <jack-mead@hotmail.co.uk>"]
license = "MIT"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
create = "src.create_provider:run"